cloudio_server/requirements.txt
absl-py==1.4.0
affine==2.4.0
astunparse==1.6.3
attrs==23.1.0
Automat==22.10.0
blinker==1.6.2
cachetools==5.3.0
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
constantly==15.1.0
construct==2.10.68
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
earthengine-api==0.1.357
Flask==2.3.2
Flask-Cors==3.0.10
flatbuffers==1.12
fonttools==4.39.4
gast==0.4.0
gcloud==0.18.3
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
imagecodecs==2023.3.16
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
Pillow==8.4.0
ply==3.11
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.18.0
pymongo==4.3.3
pyparsing==3.0.9
pyproj==3.5.0
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
rasterio==1.3.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.15.0
snuggs==1.4.7
tensor==0.3.6
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
tensorflow-metal==0.5.0
termcolor==2.3.0
threadpoolctl==3.1.0
tiff==0.1.0
tifffile==2023.4.12
Twisted==22.10.0
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.4
wrapt==1.15.0
zipp==3.15.0
zope.interface==6.0
scikit-image
